requirements.txt
Flask==2.3.2
Pillow==10.0.1
numpy==1.24.3
gunicorn==21.2.0